% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autogen.R
\name{generate_r_api}
\alias{generate_r_api}
\title{Generate an R api based on a python module}
\usage{
generate_r_api(pymodule, prefix = py_get_attr(pymodule, "__name__"))
}
\arguments{
\item{pymodule}{A python module, imported via reticulate}

\item{prefix}{prefix to give all functions and methods for generated api}
}
\value{
A string with function and method definitions
}
\description{
Generate an R api based on a python module
}
